split wallet.cpp
